divided into .c and .h
