+++ include/hw1/err.h
@@ -1,0 +1,13 @@
+#ifndef _ERR_H_
+#define _ERR_H_
+
+#define ERR_ALLOC (1)
+#define ERR_NAN (2)
+#define ERR_BIG_INPUT (3)
+#define ERR_BIG_NUMBER (4)
+#define ERR_ARR_EMPTY (5)
+#define MAX_INT (999999)
+
+int no_errors( array_return );
+
+#endif
