+++ include/hw1/numfreq.h
@@ -1,0 +1,6 @@
+#ifndef _NUMFREQ_H_
+#define _NUMFREQ_H_
+
+array_return number_frequency( array_return );
+
+#endif
